Add NuGet restore step to Publish job in pipeline
Added a new step to the `Publish` job within the `PublishAndDeploy`
stage in the `azure-pipelines.yml` file. This new step restores
NuGet packages for the `L4D2PlayStats.Web` project before the
existing step that publishes the project. This ensures all
necessary packages are restored, which is essential for the
successful publishing of the project.