Add logo image to navbar and new CSS classes
Added a new CSS class `.logo` to set the height to 100% and a maximum height of 50px. Introduced `.vertical-align-middle` class to set vertical alignment to middle with `!important` priority. Replaced the text-based navbar brand link with an image-based link, using the new `.logo` and `img-responsive` classes, and set the `alt` attribute to the localized server name.