Simplify theme logic by hardcoding `light` to false
The code change modifies the logic for determining the `light` variable. Previously, the `light` variable was set based on the value of a cookie named "theme" (it was set to `true` if the cookie's value was "light"). The new code sets the `light` variable to `false` unconditionally, removing the dependency on the cookie value.