Add onerror attribute to img tags for fallback image
Added an `onerror` attribute to various `<img>` tags in multiple `.cshtml` files to call the `imgOnError(this)` function when an image fails to load. The `imgOnError` function, defined in `_Layout.cshtml`, sets the `src` attribute to a default placeholder image (`/imgs/avatar-empty.png`). This change ensures a consistent and visually appealing interface by displaying a fallback image instead of a broken image icon.